Improve code organization some
