+++ package.json
@@ -55,7 +55,7 @@
     "source-map-support": "^0.4.6",
     "supertest": "^1.2.0",
     "tmp": "0.0.28",
-    "typescript": "^2.0.3",
+    "typescript": "^2.1.1",
     "watchy": "^0.6.6"
   },
   "typings": "./lib/polyserve.d.ts"
